+++ packages/storybook/src/data/Companies.json
@@ -1,0 +1,30 @@
+[
+  {
+    "id": 1,
+    "company": "Bradtke LLC",
+    "email": "smccunn0@symantec.com",
+    "card": "4913389273457809",
+    "country": "China"
+  },
+  {
+    "id": 2,
+    "company": "Jacobi Inc",
+    "email": "rrivard1@abc.net.au",
+    "card": "30460843005398",
+    "country": "Zimbabwe"
+  },
+  {
+    "id": 3,
+    "company": "Mraz, Blanda and Hettinger",
+    "email": "rrummins2@fastcompany.com",
+    "card": "6709016072353592810",
+    "country": "Iran"
+  },
+  {
+    "id": 4,
+    "company": "Rutherford, King and Bergstrom",
+    "email": "dkrauss3@mit.edu",
+    "card": "374288311284433",
+    "country": "Bosnia and Herzegovina"
+  }
+]
