+++ packages/storybook/src/data/Movies.json
@@ -1,0 +1,82 @@
+[
+  {
+    "id": 1,
+    "movie": "Breakout",
+    "genre": "Action|Adventure",
+    "date": "04/09/1979",
+    "country": "China",
+    "gross": "$97877758344.93"
+  },
+  {
+    "id": 2,
+    "movie": "Ladykillers, The",
+    "genre": "Comedy|Crime",
+    "date": "05/17/2018",
+    "country": "China",
+    "gross": "$9759970621.35"
+  },
+  {
+    "id": 3,
+    "movie": "Wilson",
+    "genre": "Drama",
+    "date": "01/05/1981",
+    "country": "Indonesia",
+    "gross": "$32929566919.13"
+  },
+  {
+    "id": 4,
+    "movie": "Macon County Line",
+    "genre": "Action|Drama",
+    "date": "10/26/1935",
+    "country": "Philippines",
+    "gross": "$32422166052.64"
+  },
+  {
+    "id": 5,
+    "movie": "Bless Me, Ultima",
+    "genre": "Drama|War",
+    "date": "04/01/1942",
+    "country": "Israel",
+    "gross": "$25860349604.38"
+  },
+  {
+    "id": 6,
+    "movie": "Julius Caesar",
+    "genre": "Drama",
+    "date": "12/15/1980",
+    "country": "China",
+    "gross": "$96267131975.97"
+  },
+  {
+    "id": 7,
+    "movie": "Left-Hand Side of the Fridge, The (Moitié gauche du frigo, La)",
+    "genre": "Comedy",
+    "date": "09/14/1970",
+    "country": "China",
+    "gross": "$86302448886.10"
+  },
+  {
+    "id": 8,
+    "movie": "London Conspiracy",
+    "genre": "Action|Mystery",
+    "date": "09/30/1995",
+    "country": "Norway",
+    "gross": "$70230233914.51"
+  },
+  {
+    "id": 9,
+    "movie": "Cat Run 2",
+    "genre": "Action",
+    "date": "04/10/1923",
+    "country": "Latvia",
+    "gross": "$64514647938.68"
+  },
+  {
+    "id": 10,
+    "movie": "Lady in the Water",
+    "genre": "Drama|Fantasy|Mystery",
+    "date": "12/07/1919",
+    "country": "China",
+    "gross": "$90456213202.89"
+  }
+]
