+++ packages/storybook/src/semantic-ui/EditModal.modules.js
@@ -1,10 +1,10 @@
 // @flow
 
-import React from 'react';
+import React, { useCallback, useState } from 'react';
 import { action } from '@storybook/addon-actions';
-import { withA11y } from '@storybook/addon-a11y';
-import { boolean, withKnobs } from '@storybook/addon-knobs';
+import { boolean } from '@storybook/addon-knobs';
 import {
+  Button,
   Form,
   Icon,
   Message,
@@ -15,27 +15,47 @@ import AddModal from '../components/AddModal';
 import Api from '../services/Api';
 import AssociatedDropdown from '../../../semantic-ui/src/components/AssociatedDropdown';
 import EditModal from '../../../semantic-ui/src/components/EditModal';
+import items from '../data/Companies.json';
 
 import type { EditContainerProps } from '../../../shared/src/components/EditContainer';
 
-export default {
-  title: 'Components/Semantic UI/EditModal',
-  decorators: [withA11y, withKnobs]
+const withWrapper = (WrappedComponent) => (props) => {
+  const [modal, setModal] = useState(false);
+
+  const onClose = useCallback(() => {
+    setModal(false);
+    action('close')();
+  }, []);
+
+  const onSave = useCallback(() => {
+    setModal(false);
+    action('save')();
+    return Promise.resolve();
+  }, []);
+
+  return (
+    <>
+      <Button
+        content='Open'
+        primary
+        onClick={() => setModal(true)}
+      />
+      <WrappedComponent
+        {...props}
+        onClose={onClose}
+        onSave={onSave}
+        open={modal}
+      />
+    </>
+  );
 };
 
-const onClose = action('close');
-
-const onSave = () => {
-  action('save')();
-  return Promise.resolve();
-};
-
-export const Default = () => (
+export const Default = withWrapper((props) => (
   <EditModal
-    component={(props: EditContainerProps) => (
+    component={(innerProps: EditContainerProps) => (
       <Modal
         as={Form}
-        open
+        open={props.open}
       >
         <Modal.Header
           content='Add Record'
@@ -51,46 +71,93 @@ export const Default = () => (
             label='Phone'
           />
         </Modal.Content>
-        { props.children }
+        { innerProps.children }
       </Modal>
     )}
-    onClose={onClose}
-    onSave={onSave}
+    onClose={props.onClose}
+    onSave={props.onSave}
   />
-);
+));
 
-const items = [{
-  id: 1,
-  company: 'Bradtke LLC',
-  email: 'smccunn0@symantec.com',
-  card: '4913389273457809',
-  country: 'China'
-}, {
-  id: 2,
-  company: 'Jacobi Inc',
-  email: 'rrivard1@abc.net.au',
-  card: '30460843005398',
-  country: 'Zimbabwe'
-}, {
-  id: 3,
-  company: 'Mraz, Blanda and Hettinger',
-  email: 'rrummins2@fastcompany.com',
-  card: '6709016072353592810',
-  country: 'Iran'
-}, {
-  id: 4,
-  company: 'Rutherford, King and Bergstrom',
-  email: 'dkrauss3@mit.edu',
-  card: '374288311284433',
-  country: 'Bosnia and Herzegovina'
-}];
+export const Loading = withWrapper((props) => (
+  <EditModal
+    component={(innerProps: EditContainerProps) => (
+      <Modal
+        as={Form}
+        open={props.open}
+      >
+        <Modal.Header
+          content='Add Record'
+        />
+        <Modal.Content>
+          <Form.Input
+            label='Name'
+          />
+          <Form.Input
+            label='Address'
+          />
+          <Form.Input
+            label='Phone'
+          />
+        </Modal.Content>
+        { innerProps.children }
+      </Modal>
+    )}
+    item={{ id: 1 }}
+    onClose={props.onClose}
+    onInitialize={() => new Promise((resolve) => {
+      _.delay(resolve.bind(this, {}), 5000);
+    })}
+    onSave={props.onSave}
+  />
+));
 
-export const WithToaster = () => (
+export const SaveError = withWrapper((wrapperProps) => (
   <EditModal
     component={(props: EditContainerProps) => (
       <Modal
         as={Form}
-        open
+        open={wrapperProps.open}
+        noValidate
+      >
+        <Modal.Header
+          content='Add Record'
+        />
+        <Modal.Content>
+          <Form.Input
+            required={props.isRequired('name')}
+            error={props.isError('name')}
+            label='Name'
+            onChange={props.onTextInputChange.bind(this, 'name')}
+          />
+          <Form.Input
+            required={props.isRequired('address')}
+            error={props.isError('address')}
+            label='Address'
+            onChange={props.onTextInputChange.bind(this, 'address')}
+          />
+          <Form.Input
+            required={props.isRequired('phone')}
+            error={props.isError('phone')}
+            onChange={props.onTextInputChange.bind(this, 'phone')}
+            label='Phone'
+          />
+        </Modal.Content>
+        { props.children }
+      </Modal>
+    )}
+    onClose={wrapperProps.onClose}
+    onSave={wrapperProps.onSave}
+    required={['phone', 'name', 'address']}
+  />
+));
+
+export const Toaster = withWrapper((props) => (
+  <EditModal
+    component={(innerProps: EditContainerProps) => (
+      <Modal
+        as={Form}
+        open={props.open}
       >
         <Modal.Header
           content='Add Record'
@@ -130,93 +197,20 @@ export const WithToaster = () => (
             />
           </Form.Input>
         </Modal.Content>
-        { props.children }
+        { innerProps.children }
       </Modal>
     )}
-    onClose={onClose}
-    onSave={onSave}
+    onClose={props.onClose}
+    onSave={props.onSave}
   />
-);
+));
 
-export const WithSaveError = () => (
+export const UnsavedChanges = withWrapper((wrapperProps) => (
   <EditModal
     component={(props: EditContainerProps) => (
       <Modal
         as={Form}
-        open
-        noValidate
-      >
-        <Modal.Header
-          content='Add Record'
-        />
-        <Modal.Content>
-          <Form.Input
-            required={props.isRequired('name')}
-            error={props.isError('name')}
-            label='Name'
-            onChange={props.onTextInputChange.bind(this, 'name')}
-          />
-          <Form.Input
-            required={props.isRequired('address')}
-            error={props.isError('address')}
-            label='Address'
-            onChange={props.onTextInputChange.bind(this, 'address')}
-          />
-          <Form.Input
-            required={props.isRequired('phone')}
-            error={props.isError('phone')}
-            onChange={props.onTextInputChange.bind(this, 'phone')}
-            label='Phone'
-          />
-        </Modal.Content>
-        { props.children }
-      </Modal>
-    )}
-    onClose={onClose}
-    onSave={onSave}
-    required={['phone', 'name', 'address']}
-  />
-);
-
-export const WithLoading = () => (
-  <EditModal
-    component={(props: EditContainerProps) => (
-      <Modal
-        as={Form}
-        open
-      >
-        <Modal.Header
-          content='Add Record'
-        />
-        <Modal.Content>
-          <Form.Input
-            label='Name'
-          />
-          <Form.Input
-            label='Address'
-          />
-          <Form.Input
-            label='Phone'
-          />
-        </Modal.Content>
-        { props.children }
-      </Modal>
-    )}
-    item={{ id: 1 }}
-    onClose={onClose}
-    onInitialize={() => new Promise((resolve) => {
-      _.delay(resolve.bind(this, {}), 5000);
-    })}
-    onSave={onSave}
-  />
-);
-
-export const WithUnsavedChanged = () => (
-  <EditModal
-    component={(props: EditContainerProps) => (
-      <Modal
-        as={Form}
-        open
+        open={wrapperProps.open}
       >
         <Modal.Header
           content='Add record'
@@ -259,7 +253,7 @@ export const WithUnsavedChanged = () => (
       address: '123 Main St',
       phone: null
     }}
-    onClose={onClose}
-    onSave={onSave}
+    onClose={wrapperProps.onClose}
+    onSave={wrapperProps.onSave}
   />
-);
+));
