+++ packages/storybook/src/semantic-ui/EditModal.stories.mdx
@@ -1,0 +1,75 @@
+import { Canvas, Meta, Story } from '@storybook/addon-docs';
+import DocHeader from '../components/DocHeader';
+import EditModal from '../../../semantic-ui/src/components/EditModal';
+import PropsTable from '../components/PropsTable';
+import {
+  Default,
+  Loading,
+  SaveError,
+  Toaster,
+  UnsavedChanges
+} from './EditModal.modules';
+
+<Meta
+  title='Components/Semantic UI/EditModal'
+  component={EditModal}
+/>
+
+<DocHeader
+  item={EditModal}
+/>
+
+<Canvas
+  withSource={false}
+>
+  <Story
+    name='Default'
+  >
+    <Default />
+  </Story>
+</Canvas>
+
+# Props
+
+<PropsTable
+  item={EditModal}
+/>
+
+# Examples
+#
+
+## Loading
+The `EditModal` can display with a loading indicator. This is useful when fetching data from a remote source.
+
+<Story
+  name='Loading'
+>
+  <Loading />
+</Story>
+
+## Save Error
+If the `required` prop is provided, the `EditModal` will display field validation errors when left empty.
+
+<Story
+  name='Save Error'
+>
+  <SaveError />
+</Story>
+
+## Toaster
+Use the add button on the `AssociatedDropdown` component to view the save toaster within the modal.
+
+<Story
+  name='Toaster'
+>
+  <Toaster />
+</Story>
+
+## Unsaved Changes
+The `EditModal` component will provide a `dirty` prop, which can be used to see if the record has been changed from its original state.
+
+<Story
+  name='Unsaved Changes'
+>
+  <UnsavedChanges />
+</Story>
