+++ packages/storybook/src/semantic-ui/EmbeddedList.modules.js
@@ -1,9 +1,8 @@
 // @flow
 
 import React, { useState } from 'react';
-import { withA11y } from '@storybook/addon-a11y';
 import { action } from '@storybook/addon-actions';
-import { withKnobs, optionsKnob as options } from '@storybook/addon-knobs';
+import { optionsKnob as options } from '@storybook/addon-knobs';
 import { Button, Container, Icon } from 'semantic-ui-react';
 import _ from 'underscore';
 import AddModal from '../components/AddModal';
@@ -12,11 +11,7 @@ import EditModal from '../../../semantic-ui/src/components/EditModal';
 import EmbeddedList, { SORT_DESCENDING } from '../../../semantic-ui/src/components/EmbeddedList';
 import TabbedModal from '../../../semantic-ui/src/components/TabbedModal';
 import useDragDrop from '../../../shared/src/utils/DragDrop';
-
-export default {
-  title: 'Components/Semantic UI/EmbeddedList',
-  decorators: [withA11y, withKnobs]
-};
+import items from '../data/Movies.json';
 
 const actions = [{
   name: 'edit'
@@ -43,116 +38,12 @@ const columns = [{
   label: 'Gross'
 }];
 
-const items = [{
-  id: 1,
-  movie: 'Breakout',
-  genre: 'Action|Adventure',
-  date: '04/09/1979',
-  country: 'China',
-  gross: '$97877758344.93'
-}, {
-  id: 2,
-  movie: 'Ladykillers, The',
-  genre: 'Comedy|Crime',
-  date: '05/17/2018',
-  country: 'China',
-  gross: '$9759970621.35'
-}, {
-  id: 3,
-  movie: 'Wilson',
-  genre: 'Drama',
-  date: '01/05/1981',
-  country: 'Indonesia',
-  gross: '$32929566919.13'
-}, {
-  id: 4,
-  movie: 'Macon County Line',
-  genre: 'Action|Drama',
-  date: '10/26/1935',
-  country: 'Philippines',
-  gross: '$32422166052.64'
-}, {
-  id: 5,
-  movie: 'Bless Me, Ultima',
-  genre: 'Drama|War',
-  date: '04/01/1942',
-  country: 'Israel',
-  gross: '$25860349604.38'
-}, {
-  id: 6,
-  movie: 'Julius Caesar',
-  genre: 'Drama',
-  date: '12/15/1980',
-  country: 'China',
-  gross: '$96267131975.97'
-}, {
-  id: 7,
-  movie: 'Left-Hand Side of the Fridge, The (Moitié gauche du frigo, La)',
-  genre: 'Comedy',
-  date: '09/14/1970',
-  country: 'China',
-  gross: '$86302448886.10'
-}, {
-  id: 8,
-  movie: 'London Conspiracy',
-  genre: 'Action|Mystery',
-  date: '09/30/1995',
-  country: 'Norway',
-  gross: '$70230233914.51'
-}, {
-  id: 9,
-  movie: 'Cat Run 2',
-  genre: 'Action',
-  date: '04/10/1923',
-  country: 'Latvia',
-  gross: '$64514647938.68'
-}, {
-  id: 10,
-  movie: 'Lady in the Water',
-  genre: 'Drama|Fantasy|Mystery',
-  date: '12/07/1919',
-  country: 'China',
-  gross: '$90456213202.89'
-}];
-
 export const Default = useDragDrop(() => (
   <EmbeddedList
     actions={actions}
     onDelete={action('delete')}
     columns={columns}
     items={items}
-  />
-));
-
-export const WithRecordCount = useDragDrop(() => (
-  <EmbeddedList
-    actions={actions}
-    onDelete={action('delete')}
-    columns={columns}
-    items={items}
-    showRecordCount
-  />
-));
-
-export const Selectable = useDragDrop(() => (
-  <EmbeddedList
-    selectable
-    isRowSelected={() => false}
-    onRowSelect={action('row selected')}
-    selectedRows={[{ id: 5 }, { id: 6 }, { id: 7 }]}
-    actions={actions}
-    onDelete={action('delete')}
-    columns={columns}
-    items={items}
-  />
-));
-
-export const Empty = useDragDrop(() => (
-  <EmbeddedList
-    actions={actions}
-    onDelete={action('delete')}
-    columns={columns}
-    items={[]}
   />
 ));
 
@@ -172,19 +63,14 @@ export const AddButton = useDragDrop(() => (
   />
 ));
 
-export const EmptyAddButton = useDragDrop(() => (
+export const DefaultSort = useDragDrop(() => (
   <EmbeddedList
     actions={actions}
-    addButton={{
-      location: options('Location', { Bottom: 'bottom', Top: 'top' }, 'top', { display: 'inline-radio' }),
-      color: options('Colors', SemanticColors, 'gray', { display: 'inline-radio' })
-    }}
-    onDelete={action('delete')}
     columns={columns}
-    items={[]}
-    modal={{
-      component: AddModal
-    }}
+    defaultSort='country'
+    defaultSortDirection={SORT_DESCENDING}
+    onDelete={action('delete')}
+    items={items}
   />
 ));
 
@@ -219,6 +105,63 @@ export const DragAndDropRows = useDragDrop(() => {
     />
   );
 });
+
+export const Empty = useDragDrop(() => (
+  <EmbeddedList
+    actions={actions}
+    onDelete={action('delete')}
+    columns={columns}
+    items={[]}
+  />
+));
+
+export const EmptyAddButton = useDragDrop(() => (
+  <EmbeddedList
+    actions={actions}
+    addButton={{
+      location: options('Location', { Bottom: 'bottom', Top: 'top' }, 'top', { display: 'inline-radio' }),
+      color: options('Colors', SemanticColors, 'gray', { display: 'inline-radio' })
+    }}
+    onDelete={action('delete')}
+    columns={columns}
+    items={[]}
+    modal={{
+      component: AddModal
+    }}
+  />
+));
+
+export const NoSortableColumns = useDragDrop(() => (
+  <EmbeddedList
+    actions={actions}
+    onDelete={action('delete')}
+    columns={_.map(columns, (c) => ({ ...c, sortable: false }))}
+    items={items}
+  />
+));
+
+export const RecordCount = useDragDrop(() => (
+  <EmbeddedList
+    actions={actions}
+    onDelete={action('delete')}
+    columns={columns}
+    items={items}
+    showRecordCount
+  />
+));
+
+export const Selectable = useDragDrop(() => (
+  <EmbeddedList
+    selectable
+    isRowSelected={() => false}
+    onRowSelect={action('row selected')}
+    selectedRows={[{ id: 5 }, { id: 6 }, { id: 7 }]}
+    actions={actions}
+    onDelete={action('delete')}
+    columns={columns}
+    items={items}
+  />
+));
 
 type EmbeddedListModalProps = {
   open: boolean,
@@ -280,23 +223,3 @@ export const TabbedModalConfig = useDragDrop(() => {
     </Container>
   );
 });
-
-export const NoSortableColumns = useDragDrop(() => (
-  <EmbeddedList
-    actions={actions}
-    onDelete={action('delete')}
-    columns={_.map(columns, (c) => ({ ...c, sortable: false }))}
-    items={items}
-  />
-));
-
-export const DefaultSort = useDragDrop(() => (
-  <EmbeddedList
-    actions={actions}
-    columns={columns}
-    defaultSort='country'
-    defaultSortDirection={SORT_DESCENDING}
-    onDelete={action('delete')}
-    items={items}
-  />
-));
