+++ packages/storybook/src/semantic-ui/EmbeddedList.stories.mdx
@@ -1,0 +1,126 @@
+import { Canvas, Meta, Story } from '@storybook/addon-docs';
+import EmbeddedList from '../../../semantic-ui/src/components/EmbeddedList';
+import DocHeader from '../components/DocHeader';
+import PropsTable from '../components/PropsTable';
+import {
+  Default,
+  AddButton,
+  DefaultSort,
+  DragAndDropRows,
+  Empty,
+  EmptyAddButton,
+  NoSortableColumns,
+  RecordCount,
+  Selectable,
+  TabbedModalConfig
+} from './EmbeddedList.modules';
+
+<Meta
+  title='Components/Semantic UI/EmbeddedList'
+  component={EmbeddedList}
+/>
+
+<DocHeader
+  item={EmbeddedList}
+/>
+
+<Canvas
+  withSource={false}
+>
+  <Story
+    name='Default'
+  >
+    <Default />
+  </Story>
+</Canvas>
+
+# Props
+
+<PropsTable
+  item={EmbeddedList}
+/>
+
+# Examples
+#
+
+## Add Button
+The `EmbeddedList` component can display with an add button for easily creating new records.
+
+<Story
+  name='Add Button'
+>
+  <AddButton />
+</Story>
+
+## Default Sort
+If provided, the list will be sorted by the specified column when items are provided.
+
+<Story
+  name='Default Sort'
+>
+  <DefaultSort />
+</Story>
+
+## Drag and Drop Rows
+When a `onDrag` callback is provided, rows of the table can be draggable.
+
+<Story
+  name='Drag and Drop'
+>
+  <DragAndDropRows />
+</Story>
+
+## Empty
+An empty list will display a simple message.
+
+<Story
+  name='Empty'
+>
+  <Empty />
+</Story>
+
+## Empty Add Button
+When an add button is present, the add button will become part of the empty message.
+
+<Story
+  name='Empty Add Button'
+>
+  <EmptyAddButton />
+</Story>
+
+## No Sortable Columns
+Column sorting can be disabled. This often makes sense to use in conjunction with drag and drop rows.
+
+<Story
+  name='No Sortable Columns'
+>
+  <NoSortableColumns />
+</Story>
+
+## Record Count
+The list can be configured to display the total number of records. This can be useful if using pagination or infinite scrolling.
+
+<Story
+  name='Record Count'
+>
+  <RecordCount />
+</Story>
+
+
+## Selectable
+If selectable, a checkbox will render for each row in the list.
+
+<Story
+  name='Selectable'
+>
+  <Selectable />
+</Story>
+
+## Tabbed Modal
+This example is used to test a case where we want to display the `EmbeddedList` within a `TabbedModal`.
+
+<Story
+  name='TabbedModal'
+>
+  <TabbedModalConfig />
+</Story>
