+++ packages/storybook/src/semantic-ui/FileUpload.stories.mdx
@@ -21,9 +21,14 @@ import PropsTable from '../components/PropsTable';
   </Story>
 </Canvas>
 
+# Props
+
 <PropsTable
   item={FileUpload}
 />
+
+# Examples
+#
 
 ### Size
 Using the `maxSize` prop, file sizes and be limited to exclude files exceeding a certain size.
