+++ packages/storybook/src/semantic-ui/ListTable.stories.mdx
@@ -35,7 +35,9 @@ import {
 <Canvas
   withSource={false}
 >
-  <Story name='Default'>
+  <Story
+    name='Default'
+  >
     <Default />
   </Story>
 </Canvas>
@@ -130,7 +132,7 @@ The `ListTable` can configure a custom filter modal tailored to the data being d
 </Story>
 
 ## Loading
-The `ListTable` component will display a loading indicator to notifiy the user it is fetching data.
+The `ListTable` component will display a loading indicator to notify the user it is fetching data.
 
 <Story
   name='Loading'
